some fixes for offsets

+++ grRfCom/MinIMU9AHRS/sensor.h
@@ -6,13 +6,13 @@ struct sensor {
   int AN[6];
   int AN_OFFSET[6]={0, 0, 0, 0, 0, 0};
 
-  int gyro_x;
-  int gyro_y;
-  int gyro_z;
-  int accel_x;
-  int accel_y;
-  int accel_z;
-  int magnetom_x;
-  int magnetom_y;
-  int magnetom_z;
+  float gyro_x;
+  float gyro_y;
+  float gyro_z;
+  float accel_x;
+  float accel_y;
+  float accel_z;
+  float magnetom_x;
+  float magnetom_y;
+  float magnetom_z;
 };
